implemented validation on graphql

+++ src/graphql/Mutations/Channel.ts
@@ -2,6 +2,7 @@ import { GraphQLInt, GraphQLString } from "graphql";
 import { channelServices } from "../../services/channelServices";
 import {Channel} from "../../entities/Channel.entity"
 import { channelType } from "../TypeDefs/Channel";
+import { permissions } from "../../interfaces.td";
 
 
 
@@ -15,8 +16,12 @@ export const CREATE_CHANNEL = {
         pack:{type:GraphQLInt},
     },
 
-    async resolve(parent:any ,args:Channel){
-      return await channelServices.createChannel(args)
+    async resolve(parent:any ,args:Channel, context:permissions){
+
+        if(context.role ===2)
+        return await channelServices.createChannel(args)
+
+        return []
     }
 }
 
@@ -26,8 +31,14 @@ export const DELETE_CHANNEL = {
         id:{type:GraphQLInt}
     },
 
-   async resolve(parents:any ,args:Channel){
+   async resolve(parents:any ,args:Channel, context:permissions){
+
+       if(context.role === 2 && context.isLogedin)
+       {
         await channelServices.deleteChannel(args.id)
         return "channel deleted successfully"
+       }
+
+       return " user not authorized"
     }
 }
