+++ src/graphql/Mutations/Package.ts
@@ -2,6 +2,7 @@ import {GraphQLInt, GraphQLString, GraphQLList } from "graphql";
 import { packageServices } from "../../services/packageServices";
 import {Package} from "../../entities/Package.entity"
 import { packageType } from "../TypeDefs/Package";
+import { permissions } from "../../interfaces.td";
 
 
 export const CREATE_PACKAGE = {
@@ -13,8 +14,11 @@ export const CREATE_PACKAGE = {
         price:{type:GraphQLInt}
     },
 
-    async resolve(parent:any ,args:Package){
-       return await packageServices.createPack(args)
+    async resolve(parent:any ,args:Package, context:permissions){
+        if(context.isLogedin && context.role===2)
+        return await packageServices.createPack(args)
+
+        return []
     }
 }
 
@@ -23,8 +27,13 @@ type:GraphQLString,
 args:{
     id:{type:GraphQLInt}
 },
-async resolve(parent:any , args:Package){
-    await packageServices.deletePackage(args.id)
-    return "package deleted successfully"
+async resolve(parent:any , args:Package, context:permissions){
+    if(context.isLogedin && context.role===2)
+    {
+        await packageServices.deletePackage(args.id)
+        return "package deleted successfully"
+    }
+
+    return " you are not authorized"
 }
 }
