+++ src/graphql/Mutations/User.ts
@@ -3,6 +3,7 @@ import { userServices } from "../../services/userServices";
 import {userUtils} from '../../utils/userUtils'
 import { User} from "../../entities/User.entity"
 import { userType } from "../TypeDefs/User";
+import { permissions } from "../../interfaces.td";
 
 
 
@@ -38,12 +39,14 @@ export const DELETE_USER = {
         id:{type:GraphQLInt}
     },
 
-   async resolve(parents:any ,args:User){
-     try {
-        await userServices.deleteUser(args.id)
-     } catch (error) {
-        return "user not deleted"
+   async resolve(parents:any ,args:User, context:permissions){
+     if(context.isLogedin && context.role===1)
+     {
+       await userServices.deleteUser(args.id)
+       return "user deleted successfully"
      }
-     return "user deleted successfully"
-    }
+
+
+     return " you are not authorized"
+  }
 }
